Move query complexity calculation to seperate file
