+++ src/modules/api/plugins/index.ts
@@ -1,0 +1,1 @@
+export * from './rateLimit';
